Add Upstart job to remove Libertine scope click
The unity8-dash uses a scope installed from a click before it will use
one from a deb whether or not the deb version is newer. Unfortunately,
users who have installed the Libertine scope while it was a click will
be stuck with it when they upgrade to 16.04. This commit adds an Upstart
job which checks if the libertine-scope.ubuntu click is installed, then
removes it if it is.