Revert "Ensure we don't leak references to the DBus session bus."
Using QDBusConnection::sessionBus() does NOT leak a reference to the
connection. The connection is shared accross the process and a reference
is held inside QtDbus. Simply using ::sessionBus() should increase the
speed of the calls, as we no longer repeatedly connect to the bus.

Also, when using the library from multiple threads, repeatedly connect &
disconnect using the same name can trigger a race condition in QtDbus
which will create a use-after-free resulting in a deadlock [1].

Thus, I see no reason to do it the way it did, and decided to revert
the commit, which should both increase the speed and make the library
thread-safe.

This reverts commit b6b0944e24e8059b94ef955551d0b86b07974ccd.

Fixes https://github.com/ubports/ubuntu-touch/issues/1437

[1] This is likely a bug in Qt, but I wonder if repeatedly connecting &
disconnecting the same name (and from multiple thread) is a supported
use case.